New API working with test program.
darcs-hash:20060324132308-65a35-838fca9e9a4d520e83a7ce6f0086ec9fd23d76f0.gz

+++ libsofia-sip-ua/sresolv/sofia-resolv/sres_cache.h
@@ -60,7 +60,7 @@ enum {
 #define SRES_CACHE_TIMER_INTERVAL (SRES_CACHE_TIMER_INTERVAL)
 };
 
-sres_cache_t *sres_cache_new(size_t n);
+sres_cache_t *sres_cache_new(int n);
 
 sres_cache_t *sres_cache_ref(sres_cache_t *);
